Rename concurrent_map pkg to cmap
