+++ core/master/lease_service.go
@@ -3,7 +3,7 @@ package master
 import (
 	"errors"
 	"github.com/pyropy/dfs/core/model"
-	concurrentMap "github.com/pyropy/dfs/lib/concurrent_map"
+	"github.com/pyropy/dfs/lib/cmap"
 	"time"
 
 	"github.com/google/uuid"
@@ -16,12 +16,12 @@ var (
 
 // Manages leases
 type LeaseService struct {
-	Leases concurrentMap.Map[uuid.UUID, model.Lease]
+	Leases cmap.Map[uuid.UUID, model.Lease]
 }
 
 func NewLeaseService() *LeaseService {
 	return &LeaseService{
-		Leases: concurrentMap.NewMap[uuid.UUID, model.Lease](),
+		Leases: cmap.NewMap[uuid.UUID, model.Lease](),
 	}
 }
